refactor(driver): extract protocol_txs + build_derived_block to driver/protocol_txs.rs (refactor 2.1j)
Fifth impl-block extraction of Phase 2.1. Moves the two methods
that physically construct L2 block payloads.

- `crates/based-rollup/src/driver/protocol_txs.rs` (new, 442 lines):
  - `build_builder_protocol_txs` — assembles deploys (block 1),
    setContext, setCanonicalBridgeAddress (block 2),
    loadExecutionTable, executeIncomingCrossChainCall, and drained
    mempool transactions. Owns the `max_trigger_count` parameter
    used by the §4f rebuild path in verify.rs. Also includes
    the nonce-validation loop that catches stale pool entries after
    chain rewinds.
  - `build_derived_block` — builds a block from already-assembled
    transactions using reth's `builder_for_next_block` API. Returns
    `(BuiltBlock, ExecutionData)`.

- `crates/based-rollup/src/driver/mod.rs` (-378 lines → 3002 total):
  - `mod protocol_txs;` declaration.
  - Two method bodies removed from the main `impl Driver` block.
  - Pruned now-unused mod-level imports: `alloy_consensus::BlockHeader`,
    `alloy_rpc_types_engine::ExecutionData`, `reth_evm::*`,
    `reth_payload_primitives::PayloadTypes`, `SignedTransaction`,
    `StateProviderDatabase`, `State`, `OptionExt`.
  - `DESIRED_GAS_LIMIT`, `calc_gas_limit`, `encode_block_transactions`
    moved from the unconditional `use types::{…}` to the
    `#[cfg(test)]`-gated block (tests still reference them via
    `use super::*`).

Cumulative progress:
- mod.rs: 5509 → 3002 lines (-2507 / **-45.5%** — past the halfway mark).
- 9 sibling modules totaling 2884 lines. No file > 820 lines.

`cargo check`, `cargo clippy --all-features -- -D warnings`, and
`cargo nextest -p based-rollup` (608/608) all pass. Zero behavioral
change.

Remaining in mod.rs (~3000 lines):
- Driver struct definition (~120 lines)
- Driver::new() (~200 lines)
- step/step_sync/step_builder/step_fullnode orchestration (~1500 lines)
- flush_to_l1 (~500 lines)
- misc helpers (~200 lines: drain_preconfirmed_blocks, record_l1_*,
  sync_proposer_l1_url, recover_chain_state, run, mode, compute_gas_overbid,
  collect_reverted_user_transactions, inject_held_l2_txs,
  reinject_pending_transactions, recover_builder_l2_nonce)
- derivation() accessors (~10 lines)
- test module declaration (~3 lines)

Next extractions: flush.rs (~500 lines) and step_builder.rs (~800+ lines)
are the last two large moves. After those, mod.rs will contain ~1700 lines
(struct + new + run + step + misc helpers + fullnode/sync step).
